Split schema and configs
