+++ lua/lual/config/live_level.lua
@@ -5,6 +5,7 @@
 -- For LuaRocks installed modules or busted tests, use require("lual.*")
 local core_levels = require("lua.lual.levels")
 local schemer = require("lual.utils.schemer")
+local live_level_schema_module = require("lual.config.live_level_schema")
 
 local M = {}
 
@@ -16,15 +17,6 @@ local _last_value = nil
 local _enabled = false
 local _get_env_func = os.getenv -- Can be overridden for testing
 
--- Live level configuration schema
-local live_level_schema = {
-    fields = {
-        env_var = { type = "string", required = false },
-        check_interval = { type = "number", required = false, min = 1 },
-        enabled = { type = "boolean", required = false }
-    }
-}
-
 --- Validates the live_level configuration
 -- @param config table Configuration for live_level
 -- @param full_config table The full configuration table
@@ -34,7 +26,7 @@ local function validate(config, full_config)
         return false, "live_level must be a table"
     end
 
-    local errors = schemer.validate(config, live_level_schema)
+    local errors = schemer.validate(config, live_level_schema_module.live_level_schema)
     if errors then
         return false, errors.error
     end
