+++ lua/lual/pipelines/config.lua
@@ -1,55 +1,11 @@
 --- Pipelines Configuration Handler
 -- This module handles the 'pipelines' configuration key
 
-local core_levels = require("lual.levels")
 local component_utils = require("lual.utils.component")
 local schemer = require("lual.utils.schemer")
+local pipelines_schema_module = require("lual.pipelines.schema")
 
 local M = {}
-
--- Custom validators for complex validation
-local function validate_presenter(value)
-    return type(value) == "function" or type(value) == "table"
-end
-
-local function validate_output_element(value)
-    if type(value) == "function" then
-        return true
-    end
-    if type(value) == "table" then
-        return #value > 0 or component_utils.is_callable(value)
-    end
-    return false
-end
-
--- Pipeline schema (level enum created dynamically)
-local function get_pipeline_schema()
-    return {
-        fields = {
-            level = {
-                type = "number",
-                required = false,
-                values = schemer.enum(core_levels.get_all_levels())
-            },
-            outputs = {
-                type = "table",
-                required = true,
-                count = { 1, "*" },
-                each = { custom_validator = validate_output_element }
-            },
-            presenter = {
-                required = true,
-                custom_validator = validate_presenter
-            },
-            transformers = {
-                type = "table",
-                required = false,
-                count = { 1, "*" },
-                each = { custom_validator = validate_output_element }
-            }
-        }
-    }
-end
 
 --- Validates pipelines configuration
 function M.validate(pipelines, full_config)
@@ -58,7 +14,7 @@ function M.validate(pipelines, full_config)
     end
 
     for i, pipeline in ipairs(pipelines) do
-        local errors = schemer.validate(pipeline, get_pipeline_schema())
+        local errors = schemer.validate(pipeline, pipelines_schema_module.get_pipeline_schema())
         if errors then
             return false, string.format("pipelines[%d]: %s", i, errors.error)
         end
